--- playground/set_env.py
@@ -1,5 +1,0 @@
-import os
-
-NewEnv = os.environ.copy()
-NewEnv['SCARF_NO_ANALYTICS'] = 'true'
-NewEnv['DO_NOT_TRACK'] = 'true'
